backend/requirements.txt
# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
websockets>=12.0
python-multipart>=0.0.6

# Data Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# macOS Audio Capture
pyobjc-core>=10.0
pyobjc-framework-ScreenCaptureKit>=10.0
pyobjc-framework-CoreMedia>=10.0
pyobjc-framework-AVFoundation>=10.0
pyobjc-framework-Cocoa>=10.0

# Audio Processing
numpy>=1.26.0
soundfile>=0.12.0

# LLM Clients
openai>=1.10.0
anthropic>=0.18.0
httpx>=0.26.0
dashscope>=1.14.0

# Utilities
python-dotenv>=1.0.0
loguru>=0.7.0
aiofiles>=23.2.0

# Development
pytest>=7.4.0
pytest-asyncio>=0.21.0
black>=23.12.0
ruff>=0.1.0
